+++ src/CanvasRenderingContext2d.cc
@@ -1802,8 +1802,17 @@ NAN_GETTER(Context2d::GetShadowColor) {
 
 NAN_METHOD(Context2d::SetFillColor) {
   short ok;
-  if (!info[0]->IsString()) return;
-  String::Utf8Value str(info[0]);
+
+  Isolate *isolate = Isolate::GetCurrent();
+  v8::TryCatch try_catch(isolate);
+  MaybeLocal<String> color = info[0]->ToString(Context::New(isolate));
+
+  if (try_catch.HasCaught()) {
+    try_catch.ReThrow();
+    return;
+  }
+
+  String::Utf8Value str(color.ToLocalChecked());
   uint32_t rgba = rgba_from_string(*str, &ok);
   if (!ok) return;
   Context2d *context = Nan::ObjectWrap::Unwrap<Context2d>(info.This());
@@ -1828,8 +1837,17 @@ NAN_GETTER(Context2d::GetFillColor) {
 
 NAN_METHOD(Context2d::SetStrokeColor) {
   short ok;
-  if (!info[0]->IsString()) return;
-  String::Utf8Value str(info[0]);
+
+  Isolate *isolate = Isolate::GetCurrent();
+  v8::TryCatch try_catch(isolate);
+  MaybeLocal<String> color = info[0]->ToString(Context::New(isolate));
+
+  if (try_catch.HasCaught()) {
+    try_catch.ReThrow();
+    return;
+  }
+
+  String::Utf8Value str(color.ToLocalChecked());
   uint32_t rgba = rgba_from_string(*str, &ok);
   if (!ok) return;
   Context2d *context = Nan::ObjectWrap::Unwrap<Context2d>(info.This());